Added "no Auto Dock please" functionality. To have your station skip the addition of AutoDocks, simply add the following code near the top of one of its scripts:
  local noAutoDock = require ("mods/AutoDock/data/scripts/lib/noAutoDock")

This local variable is never used, but simply including it will execute the code inside the script, causing your station to be given a "noAutoDock = true" Entity Value which the AutoDock Migrator now checks for. If this value is found, and its value is true, the AutoDock Migrator will skip that station and not add any AutoDocks to it.